fix(supabase): admin_order_list_view SUM bigint→integer 캐스트 추가
SUM(oi.quantity)는 bigint를 반환하나 기존 뷰 컬럼 itemQuantity가
integer라 타입 불일치 에러 발생. ::integer 명시적 캐스트로 수정.

+++ supabase/schemas/90_views.sql
@@ -322,7 +322,7 @@ LEFT JOIN LATERAL (
       WHERE oi2.order_id = o.id AND oi2.item_type = 'reform'
       LIMIT 1
     ) AS reform_data,
-    SUM(oi.quantity) AS item_quantity
+    SUM(oi.quantity)::integer AS item_quantity
   FROM public.order_items oi
   WHERE oi.order_id = o.id AND oi.item_type = 'reform'
 ) ri ON o.order_type IN ('custom', 'repair');
